requirements.txt
apify>=1.7.0
google-generativeai>=0.8.0
yt-dlp>=2024.1.0
